pyproject.toml
[tool.poetry]
name = "app"
version = "0.0.0"
description = ""
authors = []
license = ""

[tool.poetry.dependencies]
python = "^3.7"
uvicorn = "^0.15.0"
fastapi = "^0.68.0"
pydantic = { version = "^1.8", extras = ["email"] }
passlib = { version = "^1.7", extras = ["bcrypt"] }
pyjwt = "^2.1"
databases = "^0.5.2"
asyncpg = "^0.24.0"
psycopg2-binary = "^2.9.1"
aiosql = "^3.3.1"
pypika = "^0.48.8"
alembic = "^1.7"
python-slugify = "^5.0"
Unidecode = "^1.3"
loguru = "^0.5.1"
aiopg = "^1.3.3"

[tool.poetry.dev-dependencies]
black = "^21.9b0"
isort = "^5.9"
autoflake = "^1.4"
wemake-python-styleguide = "^0.15.3"
mypy = "^0.910"
flake8-fixme = "^1.1"
pytest = "^6.2"
pytest-cov = "^2.12"
pytest-asyncio = "^0.15.1"
pytest-env = "^0.6.2"
pytest-xdist = "^2.4.0"
httpx = "^0.19.0"
asgi-lifespan = "^1.0.1"

[tool.isort]
profile = "black"
src_paths = ["app", "tests"]
combine_as_imports = true

[tool.pytest.ini_options]
testpaths = "tests"
filterwarnings = "error"
addopts = '''
  --strict-markers
  --tb=short
  --cov=app
  --cov=tests
  --cov-branch
  --cov-report=term-missing
  --cov-report=html
  --cov-report=xml
  --no-cov-on-fail
  --cov-fail-under=100
  --numprocesses=auto
'''
env = [
  "SECRET_KEY=secret",
  "MAX_CONNECTIONS_COUNT=1",
  "MIN_CONNECTIONS_COUNT=1"
]

[build-system]
requires = ["poetry>=1.0"]
build-backend = "poetry.masonry.api"
